build(debug): fix compiler warnings about tv.tv_usec being "long"
and be strict for warnings

debug.c: In function 'univention_debug_init':
debug.c:82:48: error: format '%d' expects argument of type 'int', but argument 9 has type '__suseconds_t' {aka 'long int'} [-Werror=format=]
   82 |                 fprintf(univention_debug_file, "%04d-%02d-%02dT%02d:%02d:%02d.%06d%c%02ld:%02ld " fmt, \
      |                                                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
......
   89 |                         tv.tv_usec,\
      |                         ~~~~~~~~~~
      |                           |
      |                           __suseconds_t {aka long int}
debug.c:82:48: note: in definition of macro 'LOG'
   82 |                 fprintf(univention_debug_file, "%04d-%02d-%02dT%02d:%02d:%02d.%06d%c%02ld:%02ld " fmt, \
      |                                                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Bug #58627

+++ base/univention-debug/lib/debug.c
@@ -79,14 +79,14 @@ static const char *const univention_debug_level_structured_text[] = {
 	if (!univention_debug_structured) { \
 		fprintf(univention_debug_file, "%02d.%02d.%02d %02d:%02d:%02d.%03d  " fmt, tm.tm_mday, tm.tm_mon + 1, tm.tm_year - 100, tm.tm_hour,tm.tm_min, tm.tm_sec, (int)(tv.tv_usec / 1000), ##__VA_ARGS__); \
 	} else { \
-		fprintf(univention_debug_file, "%04d-%02d-%02dT%02d:%02d:%02d.%06d%c%02ld:%02ld " fmt, \
+		fprintf(univention_debug_file, "%04d-%02d-%02dT%02d:%02d:%02d.%06ld%c%02ld:%02ld " fmt, \
 			tm.tm_year + 1900,\
 			tm.tm_mon + 1,\
 			tm.tm_mday,\
 			tm.tm_hour,\
 			tm.tm_min,\
 			tm.tm_sec,\
-			tv.tv_usec,\
+			(long)tv.tv_usec,\
 			(tm.tm_gmtoff >= 0 ? '+' : '-'),\
 			labs(tm.tm_gmtoff) / 3600,\
 			(labs(tm.tm_gmtoff) % 3600) / 60, \
